+++ src/components/menu/types.ts
@@ -1,0 +1,1 @@
+export type Menu = { [key: string]: string | Menu }
